77975212/requirements.txt
python-dotenv==1.0.1
random-user-agent==1.0.1
requests==2.31.0
